Let's just check in the PGO Firefox binaries, one build step less
Originally, the PGO data was a bunch of proprietary blobs which I could not
legally redistribute.

When switching to Firefox binaries, I wanted to keep things as similar as
possible, to keep the repo download size down, and not waste disk space.

But in this year, megabytes are cheap, everything else is huge, I doubt many
people said no, asking the user is an extra step, and buildbots are often
sandboxed and can't download them. Let's just cut out the middlemen.